+++ CCSPayrollBillingSystem/Scripts/Employee.cs
@@ -5,7 +5,7 @@
         public int EmpIdPayroll { get; set; }
         public string EmpFnamePayroll { get; set; }
         public string EmpLnamePayroll { get; set; }
-        public double EmpRatePayroll { get; set; }
+        public decimal EmpRatePayroll { get; set; }
         public string EmpRankPayroll { get; set; }
     }
 }
